requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.5.1
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.5
django-cacheops==7.0.2
django-jet-reboot==1.3.7
django-rest-passwordreset==1.4.1
django-silk==5.1.0
djangorestframework==3.14.0
djangorestframework-yaml==2.0.0
drf-spectacular==0.27.2
flower==2.0.1
funcy==2.0
gprof2dot==2022.7.29
humanize==4.9.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
model-bakery==1.18.0
oauthlib==3.2.2
packaging==24.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.1
sentry-sdk==2.2.0
six==1.16.0
social-auth-app-django==5.4.1
sqlparse==0.5.0
tornado==6.4
tzdata==2024.1
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
